requirements.txt
coverage==4.5.3
Django==2.2.3
django-redis==4.10.0
djangorestframework==3.9.4
Pillow==6.1.0
psycopg2==2.8.3
pytz==2019.1
redis==3.2.1
sqlparse==0.3.0
